python/external-user-service/UserService/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
aiomysql==0.2.0
pymysql==1.1.0
cryptography>=40.0.0
pydantic[email]==2.5.0
pyjwt[crypto]==2.8.0
httpx==0.25.2
python-multipart==0.0.6
pydantic-settings==2.1.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator>=2.1.1
alembic==1.13.0
